requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3

# GUI (Optional)
streamlit==1.28.0

# Additional utilities
Pillow==10.1.0
matplotlib==3.8.0